[Refactor] - Move ObservableExtensions.kt to `:viewmodel` module
